refactor: use _compat/typing name
