+++ pyproject.toml
@@ -115,7 +115,7 @@ select = [
 ]
 extend-ignore = ["PLR", "E501", "SIM117"]
 target-version = "py37"
-typing-modules = ["scikit_build_core._compat.typing"]
+typing-modules = ["skbuild._compat.typing"]
 line-length = 120
 unfixable = ["T20", "F841"]
 exclude = []
